Update animal selection
